Implement acceptor creation for POSIX
This implementation behaves slightly differently compared to the Windows
implementation. On Windows, closing the socket causes the remote client
to exit, while the POSIX implementation leaves the client running after
closing the socket.

The testing client which reproduces this is the netcat utility, which is
available for both Windows and Linux.