Allow the accept callback to move the socket
Moving here means taking ownership ofthe socket. The callback must call
the move_socket function, which also tells the caller of the callback to
not destroy the socket, since it belongs to the client code now.

A new, client facing struct (ah_socket_accepted) was introduced to allow
the client code to exploit value semantics. Since the innards of this
struct are platform dependant, the AH_SOCKET_ACCEPTED_SIZE macro
controls the size of the opaque buffer in this struct. This struct is
statically asserted to match in size with the ah_socket struct. The size
may be adjusted using the adhoc-server_AH_SOCKET_ACCEPTED_SIZE CMake
cache variable.