Explicitly cast error codes
The return value of GetLastError() is an unsigned integer and the return
value of WSAGetLastError() is a signed integer, so because
WSAGetLastError() is called more, the error codes from GetLastError()
are cast to signed integer.

This doesn't affect the value of the error code, because this cast only
changes the sign of it.