Implement error code handling
The server implementation makes use of the ah_error_code enum and the
associated is_ah_error_code function to pass error codes that could
potentially be handled by client code to the callbacks or the caller of
server_tick.

Not all error codes are contained in the above mentioned enum and those
continue to be printed to stderr and the server returns with a false,
indicating a fatal error. The only case when they are not printed is
when the native event loop function fails in server_tick and error_code
is not set to NULL.